backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openpyxl
anthropic
python-dotenv
